Including socket to visualizer context
